chore: refactor S3 related integration test
