+++ extensions/data-plane/data-plane-aws-s3/src/main/java/org/eclipse/edc/connector/dataplane/aws/s3/S3DataSink.java
@@ -47,10 +47,7 @@ class S3DataSink extends ParallelSink {
     protected StreamResult<Object> transferParts(List<DataSource.Part> parts) {
         for (var part : parts) {
             try (var input = part.openStream()) {
-                String keyName = this.keyName;
-                if (parts.size() != 1) {
-                    keyName = this.keyName + "/" + part.name();
-                }
+                var keyName = parts.size() == 1 ? this.keyName : this.keyName + "/" + part.name();
 
                 var partNumber = 1;
                 var completedParts = new ArrayList<CompletedPart>();
