+++ extensions/data-plane/data-plane-aws-s3/src/test/java/org/eclipse/edc/connector/dataplane/aws/s3/S3DataPlaneIntegrationTest.java
@@ -18,8 +18,6 @@ package org.eclipse.edc.connector.dataplane.aws.s3;
 import org.eclipse.edc.aws.s3.S3BucketSchema;
 import org.eclipse.edc.aws.s3.testfixtures.AbstractS3Test;
 import org.eclipse.edc.aws.s3.testfixtures.annotations.AwsS3IntegrationTest;
-import org.eclipse.edc.connector.dataplane.aws.s3.arguments.S3DataPlaneIntegrationTestArgument;
-import org.eclipse.edc.connector.dataplane.aws.s3.arguments.S3DataPlaneIntegrationTestArgumentProvider;
 import org.eclipse.edc.spi.monitor.Monitor;
 import org.eclipse.edc.spi.security.Vault;
 import org.eclipse.edc.spi.types.TypeManager;
@@ -29,13 +27,18 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.TestInstance;
 import org.junit.jupiter.api.TestInstance.Lifecycle;
+import org.junit.jupiter.api.extension.ExtensionContext;
 import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.ArgumentsProvider;
 import org.junit.jupiter.params.provider.ArgumentsSource;
 import software.amazon.awssdk.core.ResponseBytes;
 import software.amazon.awssdk.services.s3.model.GetObjectResponse;
 
+import java.util.List;
 import java.util.UUID;
 import java.util.concurrent.Executors;
+import java.util.stream.Stream;
 
 import static java.lang.Integer.parseInt;
 import static java.util.concurrent.TimeUnit.SECONDS;
@@ -68,17 +71,10 @@ public class S3DataPlaneIntegrationTest extends AbstractS3Test {
         destinationClient.deleteBucket(destinationBucketName);
     }
 
-    @ParameterizedTest
+    @ParameterizedTest(name = "With prefix: {4}")
     @ArgumentsSource(S3DataPlaneIntegrationTestArgumentProvider.class)
-    void shouldCopyFromSourceToSink(S3DataPlaneIntegrationTestArgument arguments) {
-        var expectedKeyName = arguments.getExpectedKey();
-        var expectedKeys = arguments.getExpectedKeys();
-
-        var keys = arguments.getKeys();
-        var body = arguments.getBody();
-        var keyPrefix = arguments.getKeyPrefix();
-
-        keys.forEach(key -> sourceClient.putStringOnBucket(sourceBucketName, key, body));
+    void shouldCopyFromSourceToSink(String expectedKeyName, List<String> expectedKeys, List<String> keysToInitializeBucket, String body, String keyPrefix) {
+        keysToInitializeBucket.forEach(key -> sourceClient.putStringOnBucket(sourceBucketName, key, body));
 
         var vault = mock(Vault.class);
         var typeManager = new TypeManager();
@@ -119,14 +115,59 @@ public class S3DataPlaneIntegrationTest extends AbstractS3Test {
 
         assertThat(transferResult).succeedsWithin(5, SECONDS);
         expectedKeys.forEach(key -> {
-            assertThat(destinationClient.getObject(destinationBucketName, key)).succeedsWithin(5, SECONDS)
-                    .extracting(ResponseBytes::response)
-                    .extracting(GetObjectResponse::contentLength)
-                    .extracting(Long::intValue)
-                    .isEqualTo(body.length());
-            assertThat(destinationClient.getObject(destinationBucketName, expectedKeyName + ".complete")).succeedsWithin(5, SECONDS);
-            }
+                    assertThat(destinationClient.getObject(destinationBucketName, key)).succeedsWithin(5, SECONDS)
+                            .extracting(ResponseBytes::response)
+                            .extracting(GetObjectResponse::contentLength)
+                            .extracting(Long::intValue)
+                            .isEqualTo(body.length());
+                    assertThat(destinationClient.getObject(destinationBucketName, expectedKeyName + ".complete")).succeedsWithin(5, SECONDS);
+                }
         );
     }
 
+    public static class S3DataPlaneIntegrationTestArgumentProvider implements ArgumentsProvider {
+        private final String keyPrefix = UUID.randomUUID().toString();
+
+        private final String keyPrefixDelimiter = "/";
+
+        private final String body = UUID.randomUUID().toString();
+
+        private final List<String> keysForSingleFileTransfer = List.of(
+                generateRandomKey()
+        );
+
+        private final List<String> keysForMultipleFileTransfer = List.of(
+                generateRandomKeyWithPrefix(),
+                generateRandomKeyWithPrefix(),
+                generateRandomKeyWithPrefix()
+        );
+
+        @Override
+        public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
+            return Stream.of(
+                    Arguments.of(
+                            keysForSingleFileTransfer.get(0),
+                            keysForSingleFileTransfer,
+                            keysForSingleFileTransfer,
+                            body,
+                            null
+                    ),
+                    Arguments.of(
+                            keyPrefix,
+                            keysForMultipleFileTransfer.stream().map(key -> keyPrefix + keyPrefixDelimiter + key).toList(),
+                            keysForMultipleFileTransfer,
+                            body,
+                            keyPrefix
+                    )
+            );
+        }
+
+        private String generateRandomKey() {
+            return UUID.randomUUID().toString();
+        }
+
+        private String generateRandomKeyWithPrefix() {
+            return keyPrefix + keyPrefixDelimiter + generateRandomKey();
+        }
+    }
 }
